requirements.txt
blinker==1.7.0
Bootstrap-Flask==2.3.3
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-WTF==1.2.1
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pluggy==1.3.0
pytest-mock==3.12.0
python-dotenv==1.0.0
Werkzeug==3.0.1
WTForms==3.1.1
